[WINESYNC] ole32: Move apartment and activation management to a separate file.

wine commit id 04e11b587123e1068189015e3e6f2ad5c1faa2f8 by Nikolay Sivov <nsivov@codeweavers.com>
