+++ sdk/tools/winesync/ole32.cfg
@@ -2,4 +2,4 @@ directories:
   dlls/ole32: dll/win32/ole32
 files: include/wine/orpc.idl sdk/include/reactos/wine/orpc.idl
 tags:
-  wine: 89d8e13bf0d922bce2ed3fc9bab7d89c945d4408
+  wine: 04e11b587123e1068189015e3e6f2ad5c1faa2f8
